samples: matter: Unification of events, buttons and LEDs
This commit unifies events, buttons and LEDs for all Matter
samples.

Changelog in this commit:

- Separated Events to the system-specific and application-specific.
- Created EventManager module which is responsible for posting and
dispatching events.
- Created the common board configuration file (board_config.h)
that contains all board definitions, which can be overwritten in
the app_config.h file in the sample directory.
- Created the common board_const.h file that contains all constant
definitions of timeouts and LED blinking intervals.
Those definitions are the same for all samples and boards.
- Created the BoardInterface module in board_interface.c/.h files.
The BoardInterface module:
	- Initializes LEDs and buttons on the DK, taking
into account the number of LEDs and buttons on the DK.
	- Manages LEDs and buttons operations.
	- Manages Matter Identify service.
	- Manages to start Bluetooth LE Advertising using a button.
	- Exposes a button callback which can be used in the
Application code. Users have 3 or 1 available buttons and LEDs
depending on the availability on the DK.
	- Manages the factory reset service.
- Aligned Matter Lock sample to all changes.
	- Aligned Lock operations after clicking the button.
	- Aligned WiFi-Thread switching.
- Limitations:
	- Events cannot be called directly from the Timer timeout
event context.

Signed-off-by: Arkadiusz Balys <arkadiusz.balys@nordicsemi.no>